Correct import statement (#18)
We've noticed that the self-assign button stopped working.
This button is available in the topic controls dropdown when in mobile view. 

Our initial guess was that this might be related to this resent commit – 
discourse/discourse@fcc9d99, but I've tested and confirmed that the problem 
existed before that commit, so it regressed at some point earlier.

This PR fixes the problem by correcting an import statement. I'm not sure why 
the previous way of importing stopped working, but changing it as I've done solves 
the problem.